requirements.txt
requests>=2.31.0
feedparser>=6.0.11
python-dateutil>=2.9.0.post0
PyYAML>=6.0.2
